Refactoring and minor changes
+ repositories package
+ refactored repos to repo package
+ Show description
+ Show pictureLink
