+++ src/main/java/de/wi2020sebgroup1/cinema/repositories/CinemaRepository.java
@@ -1,8 +1,10 @@
-package de.wi2020sebgroup1.cinema.Entities;
+package de.wi2020sebgroup1.cinema.repositories;
 
 
 
 import org.springframework.data.repository.CrudRepository;
+
+import de.wi2020sebgroup1.cinema.entities.Cinema;
 
 
 public interface CinemaRepository extends CrudRepository<Cinema, Integer>{
