+++ src/main/java/de/wi2020sebgroup1/cinema/repositories/CinemaRoomRepository.java
@@ -1,7 +1,8 @@
-package de.wi2020sebgroup1.cinema.Entities;
+package de.wi2020sebgroup1.cinema.repositories;
 
 import org.springframework.data.repository.CrudRepository;
-import de.wi2020sebgroup1.cinema.Entities.CinemaRoom;
+
+import de.wi2020sebgroup1.cinema.entities.CinemaRoom;
 
 public interface CinemaRoomRepository extends CrudRepository<CinemaRoom, Integer>{
