+++ src/main/java/de/wi2020sebgroup1/cinema/repositories/CinemaRoomSeatingPlanRepository.java
@@ -1,7 +1,8 @@
-package de.wi2020sebgroup1.cinema.Entities;
+package de.wi2020sebgroup1.cinema.repositories;
 
 import org.springframework.data.repository.CrudRepository;
-import de.wi2020sebgroup1.cinema.Entities.CinemaRoomSeatingPlan;
+
+import de.wi2020sebgroup1.cinema.entities.CinemaRoomSeatingPlan;
 
 public interface CinemaRoomSeatingPlanRepository extends CrudRepository<CinemaRoomSeatingPlan, Integer> {
