+++ src/main/java/de/wi2020sebgroup1/cinema/repositories/CityRepository.java
@@ -1,8 +1,8 @@
-package de.wi2020sebgroup1.cinema.Entities;
+package de.wi2020sebgroup1.cinema.repositories;
 
 import org.springframework.data.repository.CrudRepository;
 
-import de.wi2020sebgroup1.cinema.Entities.City;
+import de.wi2020sebgroup1.cinema.entities.City;
 
 public interface CityRepository extends CrudRepository<City, Integer> {
