+++ src/main/java/de/wi2020sebgroup1/cinema/repositories/MovieRepository.java
@@ -1,7 +1,8 @@
-package de.wi2020sebgroup1.cinema.Entities;
+package de.wi2020sebgroup1.cinema.repositories;
 
 import org.springframework.data.repository.CrudRepository;
-import de.wi2020sebgroup1.cinema.Entities.Movie;
+
+import de.wi2020sebgroup1.cinema.entities.Movie;
 
 public interface MovieRepository extends CrudRepository<Movie, Integer> {
