+++ src/main/java/de/wi2020sebgroup1/cinema/repositories/PriceRepository.java
@@ -1,7 +1,8 @@
-package de.wi2020sebgroup1.cinema.Entities;
+package de.wi2020sebgroup1.cinema.repositories;
 
 import org.springframework.data.repository.CrudRepository;
-import de.wi2020sebgroup1.cinema.Entities.Price;
+
+import de.wi2020sebgroup1.cinema.entities.Price;
 
 public interface PriceRepository extends CrudRepository<Price, Integer> {
