+++ src/main/java/de/wi2020sebgroup1/cinema/repositories/RoleRepository.java
@@ -1,7 +1,8 @@
-package de.wi2020sebgroup1.cinema.Entities;
+package de.wi2020sebgroup1.cinema.repositories;
 
 import org.springframework.data.repository.CrudRepository;
-import de.wi2020sebgroup1.cinema.Entities.Role;
+
+import de.wi2020sebgroup1.cinema.entities.Role;
 
 public interface RoleRepository extends CrudRepository<Role, Integer> {
