+++ src/main/java/de/wi2020sebgroup1/cinema/repositories/SeatRepository.java
@@ -1,7 +1,8 @@
-package de.wi2020sebgroup1.cinema.Entities;
+package de.wi2020sebgroup1.cinema.repositories;
 
 import org.springframework.data.repository.CrudRepository;
-import de.wi2020sebgroup1.cinema.Entities.Seat;
+
+import de.wi2020sebgroup1.cinema.entities.Seat;
 
 public interface SeatRepository extends CrudRepository<Seat, Integer> {
