+++ src/main/java/de/wi2020sebgroup1/cinema/repositories/ShowRepository.java
@@ -1,7 +1,8 @@
-package de.wi2020sebgroup1.cinema.Entities;
+package de.wi2020sebgroup1.cinema.repositories;
 
 import org.springframework.data.repository.CrudRepository;
-import de.wi2020sebgroup1.cinema.Entities.Show;
+
+import de.wi2020sebgroup1.cinema.entities.Show;
 
 public interface ShowRepository extends CrudRepository<Show, Integer> {
