+++ src/main/java/de/wi2020sebgroup1/cinema/repositories/TicketRepository.java
@@ -1,7 +1,8 @@
-package de.wi2020sebgroup1.cinema.Entities;
+package de.wi2020sebgroup1.cinema.repositories;
 
 import org.springframework.data.repository.CrudRepository;
-import de.wi2020sebgroup1.cinema.Entities.Ticket;
+
+import de.wi2020sebgroup1.cinema.entities.Ticket;
 
 public interface TicketRepository extends CrudRepository<Ticket, Integer> {
