+++ src/main/java/de/wi2020sebgroup1/cinema/repositories/UserRepository.java
@@ -1,7 +1,8 @@
-package de.wi2020sebgroup1.cinema.Entities;
+package de.wi2020sebgroup1.cinema.repositories;
 
 import org.springframework.data.repository.CrudRepository;
-import de.wi2020sebgroup1.cinema.Entities.User;
+
+import de.wi2020sebgroup1.cinema.entities.User;
 
 public interface UserRepository extends CrudRepository<User, Integer>{
